refactor(admin): 動的部分のみを Suspense 境界に限定し PPR シェルを即時化
cacheComponents 下で各ページが先頭で verifySession/searchParams/DB を await し、
loading.tsx がコンテンツ領域全体を全画面スピナーにしていた。静的シェルが即時描画
されず、ページの素性(タイトル/説明)すら読み込み中は出ない状態だった。

- 各ページを「静的シェル(同期)＋コンテンツ Suspense」に分割
  - PageHeader/Breadcrumb/Heading 等の静的シェルは即時描画
  - verifySession・searchParams/params・DB クエリ・useSearchParams 系
    (SearchField/FilterSelect/ListPagination) はコンテンツコンポーネントへ集約し
    <Suspense fallback={<ContentFallback />}> で包む
  - searchParams/params は page で await せず Promise のまま渡す
- 共有フォールバック ContentFallback を追加し、route 直下の loading.tsx を削除
- 全 (authenticated) ルートが Partial Prerender (◐) になることを build で確認
- サイドバーの usePathname Suspense は cacheComponents 下で必須のため据え置き
