+++ apps/admin/src/app/(authenticated)/_components/content-fallback/index.ts
@@ -1,0 +1,1 @@
+export { ContentFallback } from './content-fallback';
