+++ apps/admin/src/app/(authenticated)/_components/index.ts
@@ -1,5 +1,6 @@
 // 各ページで横断的に使うレイアウト/一覧プリミティブのまとめ。
 // ページ側の import 数を抑え、利用箇所を一貫させる。
+export { ContentFallback } from './content-fallback';
 export { EmptyState } from './empty-state';
 export { FilterSelect } from './filter-select';
 export { ListPagination } from './list-pagination';
