+++ apps/admin/src/app/(authenticated)/page.tsx
@@ -1,14 +1,16 @@
-import { PageHeader } from '@/app/(authenticated)/_components/page-header';
-import { verifySession } from '@/shared/auth/verify-session';
+import { Suspense } from 'react';
+
+import { ContentFallback, PageHeader } from '@/app/(authenticated)/_components';
 
 import { DashboardContent } from './_components/dashboard-content/dashboard-content';
 
-export default async function Home() {
-  await verifySession();
+export default function Home() {
   return (
     <div className="flex flex-col gap-10">
       <PageHeader description="k8o サイト全体の概況" title="ダッシュボード" />
-      <DashboardContent />
+      <Suspense fallback={<ContentFallback />}>
+        <DashboardContent />
+      </Suspense>
     </div>
   );
 }
