+++ apps/admin/src/app/(authenticated)/reading-list/sources/[id]/page.tsx
@@ -1,18 +1,19 @@
-import { verifySession } from '@/shared/auth/verify-session';
+import { Suspense } from 'react';
+
+import { ContentFallback } from '@/app/(authenticated)/_components';
 
 import { EditSourceContent } from '../../_components/edit-source-content/edit-source-content';
 
-export default async function EditSourcePage({
+export default function EditSourcePage({
   params,
 }: {
   params: Promise<{ id: string }>;
 }) {
-  await verifySession();
-  const { id } = await params;
-
   return (
     <div className="flex flex-col gap-6">
-      <EditSourceContent id={id} />
+      <Suspense fallback={<ContentFallback />}>
+        <EditSourceContent params={params} />
+      </Suspense>
     </div>
   );
 }
