+++ apps/admin/src/app/(authenticated)/slides/page.tsx
@@ -1,20 +1,19 @@
-import { PageHeader } from '@/app/(authenticated)/_components';
-import { getSlides } from '@/features/slides/interface/queries';
-import { verifySession } from '@/shared/auth/verify-session';
+import { Suspense } from 'react';
 
-import { SlideTable } from './_components/slide-table';
+import { ContentFallback, PageHeader } from '@/app/(authenticated)/_components';
 
-export default async function SlidesPage() {
-  await verifySession();
-  const slides = await getSlides();
+import { SlidesContent } from './_components/slides-content/slides-content';
 
+export default function SlidesPage() {
   return (
     <div className="flex flex-col gap-10">
       <PageHeader
         description="スライドの公開状態を管理します（タイトル/本文は MDX 管理のため slug で表示）。"
         title="スライド"
       />
-      <SlideTable slides={slides} />
+      <Suspense fallback={<ContentFallback />}>
+        <SlidesContent />
+      </Suspense>
     </div>
   );
 }
